+++ metadata.rb
@@ -5,9 +5,10 @@ maintainer_email 'rackspace-cookbooks@rackspace.com'
 license          'Apache 2.0'
 description      'Installs/Configures elkstack'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '1.1.0'
+version          '2.0.0'
 
 depends 'apt'
+depends 'yum'
 depends 'build-essential'
 depends 'java'
 depends 'elasticsearch', '~> 0.3'