+++ test/unit/spec/forwarder_spec.rb
@@ -2,7 +2,7 @@
 
 require_relative 'helpers/spec_helper'
 
-describe 'elkstack::agent' do
+describe 'elkstack::forwarder' do
   let(:chef_run) do
     stub_resources
     ChefSpec::Runner.new(platform: 'ubuntu', version: '12.04') do |node|
@@ -14,23 +14,31 @@ describe 'elkstack::agent' do
   end
 
   it 'includes the _agent, elasticsearch::search_discovery, _secrets, and rsyslog::client recipes' do
-    expect(chef_run).to include_recipe('elkstack::agent')
+    expect(chef_run).to include_recipe('elkstack::forwarder')
     expect(chef_run).to include_recipe('elasticsearch::search_discovery')
     expect(chef_run).to include_recipe('elkstack::_secrets')
-    expect(chef_run).to include_recipe('rsyslog::client')
   end
 
-  it 'creates an instance, service, pattern and config for the agent' do
-    # enable includes start in LWRP in logstash cookbook
-    expect(chef_run).to enable_logstash_service('agent')
-
-    expect(chef_run).to create_logstash_instance('agent')
-    expect(chef_run).to create_logstash_pattern('agent')
-    expect(chef_run).to create_logstash_config('agent')
+  it 'create service for the forwarder' do
+    expect(chef_run).to enable_service('logstash-forwarder')
+    expect(chef_run).to start_service('logstash-forwarder')
   end
 
   it 'creates lumberjack key and certificate files' do
     expect(chef_run).to create_file('/opt/logstash/lumberjack.key')
     expect(chef_run).to create_file('/opt/logstash/lumberjack.crt')
   end
+
+  it 'creates logstash-forward configuration file' do
+    expect(chef_run).to create_file('/etc/logstash-forwarder')
+  end
+
+  it 'creates an apt repository' do
+    expect(chef_run).to add_apt_repository('logstash-forwarder')
+  end
+
+  it 'installs package for logstash forwarder' do
+    expect(chef_run).to install_package('logstash-forwarder')
+  end
+
 end