+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.0.0
+
+- Add a `forwarder.rb` recipe that installs [logstash-forwarder](https://github.com/elasticsearch/logstash-forwarder) as an alternative to logstash as an agent, including unit and integration tests.
+- Add additional tests for existing test-kitchen suites to ensure new lumberjack keypair is written to disk.
+- Fix a chefspec test issue where tests were checking for something that didn't make sense, didn't pass.
+
 # 1.1.0
 
 - Add `chef_environment` fields to be shipped to central cluster